% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct.R
\name{readCT}
\alias{readCT}
\title{Read files form  Expression suite software output}
\usage{
readCT(file, skip = 14, ...)
}
\arguments{
\item{file}{Name of the file to read including location}

\item{skip}{Number of line to skip form reading. Default is 14.}

\item{...}{additional parameter to input.}
}
\value{
Data frame. Need to develop in Expression class set.
}
\description{
Read files form  Expression suite software output
}
\details{
The input file can be in txt, csv or xls formate.
}
\examples{
sfile  <- system.file("extdata", "Ahlam-biopsies_20170920_131310.txt", package = "AsaLab")
mdata  <- readCT(file = sfile, skip = 14, header = FALSE)

}
\author{
Mohammad Tanvir Ahamed (mashranga@yahoo.com)
}
